add Transport#setExchangeTimeout & use in U2F instead of open timeout
this was a mistake to use the open timeout because the exchange timeout is another concept.
moreover, this was a bug that the TransportU2F was expecting a timeout in seconds and not in milliseconds like all other parts.
Fixes https://github.com/LedgerHQ/ledgerjs/issues/65

Technically, this commit is a breaking change because if you were using the U2F timeout, you should now use the .setExchangeTimeout instead. otherwise you will have the default of 30s. We probably will unroll a major release to be safe.

NB: some transport does not implement openTimeout / exchangeTimeout at the moment. we will incrementally improve this.