Revert "Removes call to reset after call to open (#609)"
We have identified the fix can cause more issues than it solves in some circumstances. It fixes the problem in the nominal scenario but there are cases where it actually makes the hardware device not "re open() able".

We are still working on finding better solutions but reverting this work is a safer move to avoid upgrading to a non-complete fix for this.

This reverts commit 3d2e728b53ea4143dea1e1d9482ab871452a9409.